Addition of plot customization features
I have added the ability to overlay the shock radius, a 2-D contour of shocks, a contour of the nse_c abundance, and tracer particles. The 2-d contour data is computed and appended to an auxiliary/reduced h5 file via my edit of write_xml.py. Full customization for the added plots is also enabled (such as line width, color, and line style).