Addition of Customization Features
I have added the ability to overlay the shock radius, a 2-D shock contour, a scatter plot of tracer particles, and a contour of the nse_c  abundance. An array of customization options of the aforementioned plots has also been added, such as customization of color, line style, and line thickness. Note: the 2-D shock contour requires derived data computed in my adaption of write_xml.py (these changes have not been committed). The write_xml.py program saves the derived values to an auxiliary/reduced h5 file. Currently, the shock radius line plot is riddled with discontinuities which I intend to address in the future.  I intend on adding a file to describe all of the customization features enumerated in setting_master.config to facilitate user understanding.